requirements.txt
opencv-contrib-python>=4.8.1.78
Pillow>=10.1.0
numpy<2.0.0
customtkinter>=5.2.2
